Adding check on if installer license file exists before trying to copy it to depot.
git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@3531 a9e16012-fd73-49ee-95cb-036de7809213

+++ bildfcns.sh
@@ -6782,11 +6782,15 @@ EOF
               if $depotDirOk; then
                 local installerTarget=`basename $installer .${sfx}`-${UQMAILHOST}.${sfx}
                 techo -2 "[$FUNCNAME] installerTarget = $installerTarget"
-                cmd="scp -v ${installerProduct}_license.txt ${INSTALLER_HOST}:${depotDir}/license.txt"
-                techo -2 "[$FUNCNAME] $cmd"
-                if ! $cmd 1>/dev/null 2>./bilderPostError; then
-                  techo "WARNING: [$FUNCNAME] '$cmd' failed: `cat bilderPostError`"
-                  rm -f bilderPostError
+                if test -f "${installerProduct}_license.txt"; then
+                  cmd="scp -v ${installerProduct}_license.txt ${INSTALLER_HOST}:${depotDir}/license.txt"
+                  techo -2 "[$FUNCNAME] $cmd"
+                  if ! $cmd 1>/dev/null 2>./bilderPostError; then
+                    techo "WARNING: [$FUNCNAME] '$cmd' failed: `cat bilderPostError`"
+                    rm -f bilderPostError
+                  fi
+                else
+                  techo "NOTE: [$FUNCNAME] Installer, $installer, found, but no license file found."
                 fi
                 cmd="scp -v $installer ${INSTALLER_HOST}:${depotDir}/${installerTarget}"
                 techo -2 "[$FUNCNAME] $cmd"
